UI QA Fixes
The close icon is displaying in a low resolution, which is causing pixelation. To address this issue, we would like to request that we upload a high definition image of the close icon. If you require any assistance in sourcing the image, please let me know and I will be happy to help.
We have noticed a discrepancy between the font-weight & the color of the card component and the design specifications. Can you please check?
We have identified a bug where the payment method selection feature is incomplete. While the user is able to select a payment method by clicking on the pill, the ability to deselect the method is currently missing.
Delete card flow
If the user reduced the height of the sheet lower than the minimum required, let us expand ourself again to the minimum needed
The delete icon has a weird transparency behind the “X” mark which is not according to our design standards.
Items separation line shouldn't be visible for the last item row
Item row in light mode has a wrong white background. It should be clear